+++ examples/server/utils.hpp
@@ -426,7 +426,7 @@ static inline std::vector<uint8_t> base64_decode(const std::string & encoded_str
     int j = 0;
     int in_ = 0;
 
-    int in_len = encoded_string.size();
+    int in_len = static_cast<int>(encoded_string.size());
 
     uint8_t char_array_4[4];
     uint8_t char_array_3[3];
@@ -440,7 +440,7 @@ static inline std::vector<uint8_t> base64_decode(const std::string & encoded_str
         {
             for (i = 0; i <4; i++)
             {
-                char_array_4[i] = base64_chars.find(char_array_4[i]);
+                char_array_4[i] = static_cast<uint8_t>(base64_chars.find(char_array_4[i]));
             }
 
             char_array_3[0] = ((char_array_4[0]      ) << 2) + ((char_array_4[1] & 0x30) >> 4);
@@ -464,7 +464,7 @@ static inline std::vector<uint8_t> base64_decode(const std::string & encoded_str
 
         for (j = 0; j <4; j++)
         {
-            char_array_4[j] = base64_chars.find(char_array_4[j]);
+            char_array_4[j] = static_cast<uint8_t>(base64_chars.find(char_array_4[j]));
         }
 
         char_array_3[0] = ((char_array_4[0]      ) << 2) + ((char_array_4[1] & 0x30) >> 4);
